分离job和shortcut
